@@ -4,31 +4,32 @@
     "private": true,
     "homepage": "/",
     "dependencies": {
-        "@material-ui/core": "^4.11.0",
-        "@testing-library/jest-dom": "^5.11.6",
-        "@testing-library/react": "^11.2.2",
-        "@testing-library/user-event": "^12.2.2",
-        "@types/history": "^4.7.6",
-        "@types/jest": "^26.0.15",
-        "@types/node": "^12.19.7",
-        "@types/react": "^16.14.2",
-        "@types/react-dom": "^16.9.10",
-        "@types/react-redux": "^7.1.9",
-        "@types/react-router-dom": "^5.1.5",
-        "axios": "^0.19.2",
-        "connected-react-router": "^6.8.0",
-        "mdi-react": "^7.3.0",
-        "react": "^17.0.1",
-        "react-dom": "^17.0.1",
-        "react-redux": "^7.2.0",
-        "react-router": "^5.2.0",
-        "react-router-dom": "^5.2.0",
-        "react-scripts": "4.0.1",
-        "redux": "^4.0.5",
-        "redux-devtools-extension": "^2.13.8",
-        "redux-thunk": "^2.3.0",
-        "typescript": "^4.1.2",
-        "web-vitals": "^0.2.4"
+        "@emotion/react": "^11.4.1",
+        "@emotion/styled": "^11.3.0",
+        "@mui/material": "^5.0.0",
+        "@reduxjs/toolkit": "^1.6.1",
+        "@testing-library/jest-dom": "^5.14.1",
+        "@testing-library/react": "^12.1.0",
+        "@testing-library/user-event": "^13.2.1",
+        "@types/history": "^4.7.9",
+        "@types/jest": "^27.0.2",
+        "@types/node": "^16.9.6",
+        "@types/react": "^17.0.24",
+        "@types/react-dom": "^17.0.9",
+        "@types/react-redux": "^7.1.18",
+        "@types/react-router-dom": "^5.3.0",
+        "axios": "^0.21.4",
+        "connected-react-router": "^6.9.1",
+        "mdi-react": "^7.5.0",
+        "react": "^17.0.2",
+        "react-dom": "^17.0.2",
+        "react-redux": "^7.2.5",
+        "react-router": "^5.2.1",
+        "react-router-dom": "^5.3.0",
+        "react-scripts": "4.0.3",
+        "redux": "^4.1.1",
+        "typescript": "^4.4.3",
+        "web-vitals": "^2.1.0"
     },
     "scripts": {
         "start": "react-scripts start",
@@ -55,6 +56,6 @@
         ]
     },
     "devDependencies": {
-        "node-sass": "^4.14.1"
+        "node-sass": "^6.0.1"
     }
 }
